+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Updated
 - k-fold splitter to include validation set
+- Transparent file opening for compressed files
 
 ### Fixed
 - Missing data one hot encoder to handle sparse inputs